+++ CHANGELOG.md
@@ -6,7 +6,12 @@ All notable changes to this project will be documented in this file.
 - SDK Authorization support
 
 ### New
-- `authorization` config parameter used to specify SDK access key.
+- `authorization` config parameter to specify SDK access key.
+- `registerAccessKey` customer account management function.
+- `revokeAccessKey` customer account management function.
+
+### Fixed
+- flow type definitions `types.js` have actualized.
 
 ## 0.20.0 - Feb 4, 2020